Upgrade actions from v3 to v4
Seeing alot of deprecation warnings in jobs. Updating from v3 to v4 to remove the warnings.

Checked all of the repo .tool-versions using this search path:**/.tool-versions

Only a few repos are using a node version under 20, the only one using the shared actions is the public repo https://github.com/BuoySoftware/anchor-ui, which is legacy code and not in use.

Also ensured all of the same inputs are still in use for setup-node, cache and checkout